Fixes bug with blank remember_tokens
In tutorial, it assumes user will set all users in db to have a token;
but if user did not and was already logged in, they can never log out.

This patch checks that the remember_token is not blank before using it
to find a user associated with it.